Set up a scene to do cinematics and separated it from rest of the project.
I did this by unpacking all prefabs and disabling or removing entirely most of the scripts. I didn't want to have any conflicts in the final hour. I created a separate player prefab and animators. I created a separate script with the intent to control player animation directly and so on. Also added a new timeline for creating credits cinematic.